feat: status checks for sites
Now that we are starting to use GH actions for sites, it is helpful
to be reporting on the status of each pull request. This bumps the
rate cost for the status query from 35 to 60. Together with the current
contributions query cost of 3, we should still be well within 5000/hour
(we refresh once per minute = 3780).