fix: version number extraction on pso and llso
The source package isn't always listed with a version number on
pso and llso. This change uses the `Version` field instead to
retrieve the version number. This gives the version of the binary
package - which should be identical to the version of the source
package, at least the Keyman version part.