feat: add has-user-test label
Fixes #186.

Adds a 'has-user-test' label to each PR or issue that has a `# User
Testing` section. Removes the label if that section subsequently
disappears.

Note that existing PRs and issues will not be affected by this, so we
should go ahead and manually search for those and add the label.