feat: add 'SKIPPED' test result type
Fixes #186.

'SKIPPED' test result type should be used when you want to treat a test
as passing but the test was not actually completed successfully. For
example, you may want to mark a test as 'SKIPPED' if it turns out to be
irrelevant, or if a bug elsewhere stops the test from being completed,
although in most cases you should probably use 'BLOCKED' if a test is
never completed successfully, as this more clearly indicates that we
have not actually verified that things work as we hope.