chore: manage rate limits for GitHub
Fixes #420.

Adds server-side logging for rate limiting and cost for each GraphQL
query. Note that this misses rate limit data for the REST API usage
(currently only github-milestones.ts, which should have a cost of 1).

Tweaks the expensive github-status:organization query to reduce the cost
from 100 to 48, by retrieving fewer PRs per repo (max of 20), and max
of 60 repos. This means that if we start getting many PRs on a single
repo, we may not have visibility.

This is an interim fix, would be much better to use cached data and
fold changes in that are given to us by the GH webhook notifications,
but that's a much, much more complex change.