feat: add support for `Test-bot:` trailer in PR comments
The existing `@keymanapp-test-bot` pattern will continue to be
supported, but the commit-trailer style `Test-bot: ` matches other
controls we use in commit messages and is recommended for use.

Note that this trailer is only recognized in PR comments at this time;
parsing commit messages may be supported in a future update (but with
caution because they are less visible in the PR view and workflow, so
may cause confusion to testers).

Fixes: keymanapp/keyman#13212