feat: diff-based refresh of GitHub issues
The most costly data load for status.keyman.com is the loading of open
issues, which takes upwards of 30 seconds. The current design was
reloading the entire issue list for every single issues, issue_comment,
pull_request, check_run, or check_suite webhook event -- which is
incredibly wasteful.

This change uses the webhook event type and body content to determine
which data needs to be reloaded, which firstly avoids issues being
reloaded constantly for unrelated events, and then secondly, does a
targeted load of just the impacted issue for issues and issue_comment
events, adding, replacing, or deleting just that issue from the cache.

The data sent through to clients is still a full set, not just the
change, but this can come later, as it doesn't impact our rate limit on
GitHub.

As a secondary improvement, issue_comment events now only trigger a
reload of user test contribution data if the affected comment contains
a valid user test pattern (note that if a comment edit _removes_ a
valid user test pattern form the comment, the issue will remain in
contributions, but this is a rare thing and not worth stressing over
at this point).

We should be able to do similar things in the future for other costly
GitHub data, but the issue list was the most costly, by far, of all the
events. We are pushing the boundaries in the honking big request in
github-status.ts, so it would be good to break this down and then apply
similar principles -- probably, merging all pull request data into a
single array in the cache (it has evolved and there are multiple sources
of the PR data in the cache right now), then applying the same
principles as I did with issues in this commit.