fix: enable Sentry error reporting for front-end
I noticed this when I was fixing a bug in #531, and observed that
script errors client-side were not being logged to Sentry. We will
probably start to get a lot of messages into Sentry; once we clear off
most of the chaos, it may be worth turning on the Sentry dialog in
app.module.ts.

Test-bot: skip
Follows: #532